requirements.txt
fastapi
uvicorn
requests
sqlalchemy
pydantic
slowapi
fastapi-cache2
